IntelliJ-Plugin for GLSL syntax highlighting and on-the-fly preview.
The plugin will highlight GLSL syntax (focus on GLSL ES 2.0) in IntelliJ IDEA and AndroidStudio. 
Another goal is to develop a live view of the current shader code like seen in some WebGL tools:
http://www.kickjs.org/tool/shader_editor/shader_editor.html#
It also would be great if there is some way to customize the previewed primitive and to support
user edited uniforms.